# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NFL Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/schedule": {
      "get": {
        "summary": "Get Schedule",
        "operationId": "get_schedule_schedule_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2025,
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleWeek"
                  },
                  "title": "Response Get Schedule Schedule Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games": {
      "get": {
        "summary": "Get Games",
        "operationId": "get_games_games_get",
        "parameters": [
          {
            "name": "week",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 22,
              "minimum": 1,
              "title": "Week"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2025,
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  },
                  "title": "Response Get Games Games Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}": {
      "get": {
        "summary": "Get Game",
        "operationId": "get_game_games__game_id__get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Game": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "game_type": {
            "type": "string",
            "title": "Game Type"
          },
          "week": {
            "type": "integer",
            "title": "Week"
          },
          "gameday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gameday"
          },
          "gametime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gametime"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Score"
          },
          "home_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Score"
          },
          "away_qb_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Qb Name"
          },
          "home_qb_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Qb Name"
          },
          "spread_line": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spread Line"
          },
          "total_line": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Line"
          },
          "roof": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roof"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "temp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp"
          },
          "wind": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind"
          },
          "stadium": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stadium"
          },
          "overtime": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overtime"
          },
          "div_game": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Div Game"
          },
          "away_record": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Record"
          },
          "home_record": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Record"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "season",
          "game_type",
          "week",
          "gameday",
          "gametime",
          "away_team",
          "home_team",
          "away_score",
          "home_score",
          "away_qb_name",
          "home_qb_name",
          "spread_line",
          "total_line",
          "roof",
          "surface",
          "temp",
          "wind",
          "stadium",
          "overtime",
          "div_game"
        ],
        "title": "Game",
        "description": "A scheduled game; scores are null for unplayed games."
      },
      "GameDetail": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "game_type": {
            "type": "string",
            "title": "Game Type"
          },
          "week": {
            "type": "integer",
            "title": "Week"
          },
          "gameday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gameday"
          },
          "gametime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gametime"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Score"
          },
          "home_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Score"
          },
          "away_qb_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Qb Name"
          },
          "home_qb_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Qb Name"
          },
          "spread_line": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spread Line"
          },
          "total_line": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Line"
          },
          "roof": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roof"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "temp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp"
          },
          "wind": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind"
          },
          "stadium": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stadium"
          },
          "overtime": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overtime"
          },
          "div_game": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Div Game"
          },
          "away_record": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Record"
          },
          "home_record": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Record"
          },
          "away": {
            "items": {
              "$ref": "#/components/schemas/GamePlayerStats"
            },
            "type": "array",
            "title": "Away"
          },
          "home": {
            "items": {
              "$ref": "#/components/schemas/GamePlayerStats"
            },
            "type": "array",
            "title": "Home"
          },
          "quarter_scores": {
            "items": {
              "$ref": "#/components/schemas/QuarterScore"
            },
            "type": "array",
            "title": "Quarter Scores"
          },
          "win_prob": {
            "items": {
              "$ref": "#/components/schemas/WinProbPlay"
            },
            "type": "array",
            "title": "Win Prob"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "season",
          "game_type",
          "week",
          "gameday",
          "gametime",
          "away_team",
          "home_team",
          "away_score",
          "home_score",
          "away_qb_name",
          "home_qb_name",
          "spread_line",
          "total_line",
          "roof",
          "surface",
          "temp",
          "wind",
          "stadium",
          "overtime",
          "div_game",
          "away",
          "home",
          "quarter_scores",
          "win_prob"
        ],
        "title": "GameDetail",
        "description": "Game detail returned by /games/{id}: game header + rosters + scoring + WP."
      },
      "GamePlayerStats": {
        "properties": {
          "player_id": {
            "type": "string",
            "title": "Player Id"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "team": {
            "type": "string",
            "title": "Team"
          },
          "week": {
            "type": "integer",
            "title": "Week"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "jersey_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jersey Number"
          },
          "headshot_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headshot Url"
          },
          "attempts": {
            "type": "number",
            "title": "Attempts"
          },
          "completions": {
            "type": "number",
            "title": "Completions"
          },
          "pass_yards": {
            "type": "number",
            "title": "Pass Yards"
          },
          "pass_tds": {
            "type": "number",
            "title": "Pass Tds"
          },
          "interceptions_thrown": {
            "type": "number",
            "title": "Interceptions Thrown"
          },
          "sacks_taken": {
            "type": "number",
            "title": "Sacks Taken"
          },
          "pass_epa": {
            "type": "number",
            "title": "Pass Epa"
          },
          "targets": {
            "type": "number",
            "title": "Targets"
          },
          "receptions": {
            "type": "number",
            "title": "Receptions"
          },
          "rec_yards": {
            "type": "number",
            "title": "Rec Yards"
          },
          "rec_tds": {
            "type": "number",
            "title": "Rec Tds"
          },
          "air_yards": {
            "type": "number",
            "title": "Air Yards"
          },
          "yac": {
            "type": "number",
            "title": "Yac"
          },
          "rec_epa": {
            "type": "number",
            "title": "Rec Epa"
          },
          "carries": {
            "type": "number",
            "title": "Carries"
          },
          "rush_yards": {
            "type": "number",
            "title": "Rush Yards"
          },
          "rush_tds": {
            "type": "number",
            "title": "Rush Tds"
          },
          "rush_epa": {
            "type": "number",
            "title": "Rush Epa"
          },
          "solo_tackles": {
            "type": "number",
            "title": "Solo Tackles"
          },
          "assist_tackles": {
            "type": "number",
            "title": "Assist Tackles"
          },
          "tackles_for_loss": {
            "type": "number",
            "title": "Tackles For Loss"
          },
          "sacks": {
            "type": "number",
            "title": "Sacks"
          },
          "qb_hits": {
            "type": "number",
            "title": "Qb Hits"
          },
          "def_interceptions": {
            "type": "number",
            "title": "Def Interceptions"
          },
          "pass_breakups": {
            "type": "number",
            "title": "Pass Breakups"
          },
          "forced_fumbles": {
            "type": "number",
            "title": "Forced Fumbles"
          },
          "fumble_recoveries": {
            "type": "number",
            "title": "Fumble Recoveries"
          },
          "fg_att": {
            "type": "number",
            "title": "Fg Att"
          },
          "fg_made": {
            "type": "number",
            "title": "Fg Made"
          },
          "xp_att": {
            "type": "number",
            "title": "Xp Att"
          },
          "xp_made": {
            "type": "number",
            "title": "Xp Made"
          },
          "punts": {
            "type": "number",
            "title": "Punts"
          },
          "punt_yards": {
            "type": "number",
            "title": "Punt Yards"
          },
          "punt_returns": {
            "type": "number",
            "title": "Punt Returns"
          },
          "punt_return_yards": {
            "type": "number",
            "title": "Punt Return Yards"
          },
          "punt_return_tds": {
            "type": "number",
            "title": "Punt Return Tds"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "player_name",
          "team",
          "week",
          "position",
          "jersey_number",
          "headshot_url",
          "attempts",
          "completions",
          "pass_yards",
          "pass_tds",
          "interceptions_thrown",
          "sacks_taken",
          "pass_epa",
          "targets",
          "receptions",
          "rec_yards",
          "rec_tds",
          "air_yards",
          "yac",
          "rec_epa",
          "carries",
          "rush_yards",
          "rush_tds",
          "rush_epa",
          "solo_tackles",
          "assist_tackles",
          "tackles_for_loss",
          "sacks",
          "qb_hits",
          "def_interceptions",
          "pass_breakups",
          "forced_fumbles",
          "fumble_recoveries",
          "fg_att",
          "fg_made",
          "xp_att",
          "xp_made",
          "punts",
          "punt_yards",
          "punt_returns",
          "punt_return_yards",
          "punt_return_tds"
        ],
        "title": "GamePlayerStats",
        "description": "Per-game player line returned in /games/{id} under away[] / home[].\n\nStat columns come back as DOUBLE from DuckDB (DEFAULT 0), so they are\nfloats even when conceptually counts (touchdowns, attempts, etc.)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuarterScore": {
        "properties": {
          "qtr": {
            "type": "integer",
            "title": "Qtr"
          },
          "away": {
            "type": "integer",
            "title": "Away"
          },
          "home": {
            "type": "integer",
            "title": "Home"
          }
        },
        "type": "object",
        "required": [
          "qtr",
          "away",
          "home"
        ],
        "title": "QuarterScore"
      },
...
      "ScheduleWeek": {
        "properties": {
          "week": {
            "type": "integer",
            "title": "Week"
          },
          "games": {
            "items": {
              "$ref": "#/components/schemas/Game"
            },
            "type": "array",
            "title": "Games"
          }
        },
        "type": "object",
        "required": [
          "week",
          "games"
        ],
        "title": "ScheduleWeek",
        "description": "A week of games returned by /schedule."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WinProbPlay": {
        "properties": {
          "game_seconds_remaining": {
            "type": "integer",
            "title": "Game Seconds Remaining"
          },
          "qtr": {
            "type": "integer",
            "title": "Qtr"
          },
          "home_wp": {
            "type": "number",
            "title": "Home Wp"
          },
          "touchdown": {
            "type": "integer",
            "title": "Touchdown"
          },
          "interception": {
            "type": "integer",
            "title": "Interception"
          },
          "fumble_lost": {
            "type": "integer",
            "title": "Fumble Lost"
          },
          "posteam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posteam"
          },
          "desc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desc"
          }
        },
        "type": "object",
        "required": [
          "game_seconds_remaining",
          "qtr",
          "home_wp",
          "touchdown",
          "interception",
          "fumble_lost",
          "posteam",
          "desc"
        ],
        "title": "WinProbPlay"
      },
...
    }
  }
}